Update README
Taskhive is free(libre), open source software (FOSS/FLOSS). Anyone can crack open the code and see what it does, modify it, and share it with the world. But most importantly, it can be audited and prove that it is not doing anything milacious.
I think that there is a typo in the end and must be malicious.